+++ src/common.rs
@@ -1,11 +1,11 @@
 use rand::Rng;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
-pub fn in_range(min: f64, x: f64, max: f64) -> bool {
+pub fn in_range(min: f32, x: f32, max: f32) -> bool {
     x > min && x < max
 }
 
-pub fn clamp(min: f64, x: f64, max: f64) -> f64 {
+pub fn clamp(min: f32, x: f32, max: f32) -> f32 {
     if x < min {
         return min;
     }
@@ -15,10 +15,10 @@ pub fn clamp(min: f64, x: f64, max: f64) -> f64 {
     x
 }
 
-pub fn random_in_range(min: f64, max: f64) -> f64 {
-    min + (max - min) * rand::rng().random::<f64>()
+pub fn random_in_range(min: f32, max: f32) -> f32 {
+    min + (max - min) * rand::rng().random::<f32>()
 }
 
-pub fn degrees_to_radians(deg: f64) -> f64 {
+pub fn degrees_to_radians(deg: f32) -> f32 {
     deg * PI / 180.0
 }
